requirements.txt
rich
termcolor
watchdog
PyYAML
requests
pyvmomi
